Split dummy data and definitions
